move downstream checkpoints test implementation to appropriate module
